services/frontend/requirements.txt
streamlit
requests
python-dotenv
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-google-genai>=1.0.0
chromadb
GitPython
tiktoken
